Condense timing-related structs
