+++ src/engine.rs
@@ -24,7 +24,7 @@ use sawtooth_sdk::consensus::{engine::*, service::Service};
 use node::PbftNode;
 
 use config;
-use ticker;
+use timing;
 
 // How long to wait in between trying to publish blocks
 const BLOCK_DURATION: Duration = Duration::from_millis(2000);
@@ -53,7 +53,7 @@ impl Engine for PbftEngine {
         chain_head: Block,
         _peers: Vec<PeerInfo>,
     ) {
-        let mut working_ticker = ticker::Ticker::new(BLOCK_DURATION);
+        let mut working_ticker = timing::Ticker::new(BLOCK_DURATION);
 
         let config = config::load_pbft_config(chain_head.block_id, &mut service);
