+++ src/timing.rs
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Intel Corporation
+ * Copyright 2018 Bitwise IO, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,11 +17,10 @@
 
 use std::time::{Duration, Instant};
 
-/// Encapsulates doing some work every time a timeout has elapsed
+// Encapsulates doing some work every time a timeout has elapsed
 pub struct Ticker {
     last: Instant,
     timeout: Duration,
-    remaining: Duration,
 }
 
 impl Ticker {
@@ -29,16 +28,47 @@ impl Ticker {
         Ticker {
             last: Instant::now(),
             timeout: period,
-            remaining: period,
         }
     }
 
-    /// Do some work if the timeout has elapsed
+    // Do some work if the timeout has elapsed
     pub fn tick<T: FnMut()>(&mut self, mut callback: T) {
         let elapsed = Instant::now() - self.last;
         if elapsed >= self.timeout {
             callback();
             self.last = Instant::now();
+        }
+    }
+}
+
+enum TimeoutState {
+    Active,
+    Expired,
+}
+
+// Check back on this timer every so often to see if it's expired
+pub struct Timeout {
+    state: TimeoutState,
+    duration: Duration,
+    start: Instant,
+}
+
+impl Timeout {
+    pub fn new(duration: Duration) -> Self {
+        Timeout {
+            state: TimeoutState::Active,
+            duration: duration,
+            start: Instant::now(),
+        }
+    }
+
+    pub fn is_expired(&mut self) -> bool {
+        if Instant::now() - self.start > self.duration {
+            self.state = TimeoutState::Expired;
+        }
+        match self.state {
+            TimeoutState::Active => false,
+            TimeoutState::Expired => true,
         }
     }
 }
